requirements.txt
docling
flask
pyyaml
openaik